refactor and add some logs
